LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/detached-context.https-expected.txt


PASS accessing a ServiceWorkerRegistration from a removed iframe 
PASS accessing a ServiceWorker object from a removed iframe 
PASS accessing navigator.serviceWorker on a detached iframe 
FAIL accessing navigator on a removed frame assert_throws_js: function "() => get_navigator()" did not throw
FAIL accessing navigator.serviceWorker on a removed about:blank frame assert_equals: expected null but got object "[object ServiceWorkerContainer]"
